Audio/ValKit: More code for the audio test execution service (ATS), client side [Windows build fix]. bugref:10008
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@89206 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/AudioTestServiceClient.cpp
@@ -187,7 +187,7 @@ DECLINLINE (void) audioTestSvcClientReqHdrInit(PATSPKTHDR pReqHdr, size_t cbReq,
 
     memcpy(pReqHdr->achOpcode, pszOp, sizeof(pReqHdr->achOpcode));
     pReqHdr->uCrc32 = 0; /** @todo Do CRC-32 calculation. */
-    pReqHdr->cb     = cbReq + cbPayload;
+    pReqHdr->cb     = (uint32_t)cbReq + (uint32_t)cbPayload;
 }
 
 static int audioTestSvcClientDoGreet(PATSCLIENT pClient)
